backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database & ORM
supabase
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0

# Rate Limiting & Security
slowapi==0.1.9
python-dotenv==1.0.0

# HTTP & API Clients

requests==2.31.0
aiohttp==3.9.1

# AI & Content Generation
openai==1.10.0

# LinkedIn API
linkedin-api==2.2.0
requests-oauthlib==1.3.1

# Input Validation & Sanitization
bleach==6.1.0
validators==0.22.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
APScheduler==3.10.4

# Logging & Monitoring
loguru==0.7.2

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3


# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
